fix: ignoringOnTheFly is now more accurate to the original python implementation
basically, it turns out IgnoringOnTheFly is an integer for a reason, so the client/server knows the other acknowledged the last ignore request it sent. if the server sent a 1 and then a 2 immediately after, the server expects a 2 back and will ignore a 1 response if the client sends that, assuming it's not up to date. (i think. poor explanation)